# src/main/java/org/openapitools/api/ArtistsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Album;
import org.openapitools.model.Artist;
import org.openapitools.model.CreateAlbumRequest;
import org.openapitools.model.Subscription;
import org.openapitools.model.Track;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-03T18:02:44.513930+01:00[Europe/Madrid]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "contents", description = "Operaciones de artistas, discográficas, álbumes, canciones y archivos de audio.")
public interface ArtistsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_ALBUM = "/artists/{idArtist}/albums";
    /**
     * POST /artists/{idArtist}/albums : Crear álbum
     * Crea un nuevo álbum asociado a un artista (requiere rol artista).
     *
     * @param idArtist  (required)
     * @param createAlbumRequest Datos del nuevo álbum (required)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "createAlbum",
        summary = "Crear álbum",
        description = "Crea un nuevo álbum asociado a un artista (requiere rol artista).",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created"),
            @ApiResponse(responseCode = "400", description = "Bad Request"),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ArtistsApi.PATH_CREATE_ALBUM,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> createAlbum(
        @NotNull @Parameter(name = "idArtist", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idArtist") String idArtist,
        @Parameter(name = "CreateAlbumRequest", description = "Datos del nuevo álbum", required = true) @Valid @RequestBody CreateAlbumRequest createAlbumRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_ARTIST_BY_ID = "/artists/{idArtist}";
    /**
     * GET /artists/{idArtist} : Información de artista
     * Devuelve la información detallada de un artista por su identificador.
     *
     * @param idArtist  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "getArtistById",
        summary = "Información de artista",
        description = "Devuelve la información detallada de un artista por su identificador.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Artist.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtistsApi.PATH_GET_ARTIST_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<Artist> getArtistById(
        @NotNull @Parameter(name = "idArtist", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idArtist") String idArtist
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"user_id\" : \"usr_001\", \"name\" : \"David bisbal\", \"genre\" : \"Ambient\", \"description\" : \"Artista de pop\", \"id\" : \"art_001\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_ARTIST_SUBSCRIBERS = "/artists/{idArtist}/subscriptions";
    /**
     * GET /artists/{idArtist}/subscriptions : Usuarios suscritos al artista
     * Devuelve los usuarios que siguen al artista especificado.
     *
     * @param idArtist  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "getArtistSubscribers",
        summary = "Usuarios suscritos al artista",
        description = "Devuelve los usuarios que siguen al artista especificado.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Subscription.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtistsApi.PATH_GET_ARTIST_SUBSCRIBERS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Subscription>> getArtistSubscribers(
        @NotNull @Parameter(name = "idArtist", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idArtist") String idArtist
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"user_id\" : \"usr_777\", \"created_at\" : \"2025-10-15T18:00:00Z\", \"artist_id\" : \"art_001\" }, { \"user_id\" : \"usr_777\", \"created_at\" : \"2025-10-15T18:00:00Z\", \"artist_id\" : \"art_001\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_ARTIST_ALBUMS = "/artists/{idArtist}/albums";
    /**
     * GET /artists/{idArtist}/albums : Álbumes por artista
     * Obtiene todos los álbumes publicados por un artista determinado.
     *
     * @param idArtist  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "listArtistAlbums",
        summary = "Álbumes por artista",
        description = "Obtiene todos los álbumes publicados por un artista determinado.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Album.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtistsApi.PATH_LIST_ARTIST_ALBUMS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Album>> listArtistAlbums(
        @NotNull @Parameter(name = "idArtist", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idArtist") String idArtist
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"cover_url\" : \"https://avatar.iran.liara.run/public/40\", \"release_date\" : \"2008-09-09\", \"id\" : \"alb_101\", \"title\" : \"La ley innata\", \"artist_id\" : \"art_001\" }, { \"cover_url\" : \"https://avatar.iran.liara.run/public/40\", \"release_date\" : \"2008-09-09\", \"id\" : \"alb_101\", \"title\" : \"La ley innata\", \"artist_id\" : \"art_001\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_ARTIST_TRACKS = "/artists/{idArtist}/tracks";
    /**
     * GET /artists/{idArtist}/tracks : Canciones por artista
     * Devuelve todas las canciones publicadas por un artista.
     *
     * @param idArtist  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "listArtistTracks",
        summary = "Canciones por artista",
        description = "Devuelve todas las canciones publicadas por un artista.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Track.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtistsApi.PATH_LIST_ARTIST_TRACKS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Track>> listArtistTracks(
        @NotNull @Parameter(name = "idArtist", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idArtist") String idArtist
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"duration\" : \"03:45\", \"file_url\" : \"https://cdn.musica.com/tracks/de-acero.mp3\", \"album_id\" : \"alb_101\", \"id\" : \"trk_301\", \"title\" : \"De acero\", \"published_at\" : \"2025-10-16T20:00:00Z\" }, { \"duration\" : \"03:45\", \"file_url\" : \"https://cdn.musica.com/tracks/de-acero.mp3\", \"album_id\" : \"alb_101\", \"id\" : \"trk_301\", \"title\" : \"De acero\", \"published_at\" : \"2025-10-16T20:00:00Z\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_ARTISTS = "/artists";
    /**
     * GET /artists : Listado de artistas
     * Devuelve el listado completo de artistas disponibles en la plataforma.
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "listArtists",
        summary = "Listado de artistas",
        description = "Devuelve el listado completo de artistas disponibles en la plataforma.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Artist.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ArtistsApi.PATH_LIST_ARTISTS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Artist>> listArtists(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"user_id\" : \"usr_001\", \"name\" : \"David bisbal\", \"genre\" : \"Ambient\", \"description\" : \"Artista de pop\", \"id\" : \"art_001\" }, { \"user_id\" : \"usr_001\", \"name\" : \"David bisbal\", \"genre\" : \"Ambient\", \"description\" : \"Artista de pop\", \"id\" : \"art_001\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}