# src/main/java/org/openapitools/api/TracksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Track;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-03T18:02:44.513930+01:00[Europe/Madrid]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "contents", description = "Operaciones de artistas, discográficas, álbumes, canciones y archivos de audio.")
public interface TracksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_TRACK_BY_ID = "/tracks/{idTrack}";
    /**
     * GET /tracks/{idTrack} : Información de canción por ID
     * Devuelve los datos de una canción específica.
     *
     * @param idTrack  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "getTrackById",
        summary = "Información de canción por ID",
        description = "Devuelve los datos de una canción específica.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Track.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TracksApi.PATH_GET_TRACK_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<Track> getTrackById(
        @NotNull @Parameter(name = "idTrack", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idTrack") String idTrack
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"duration\" : \"03:45\", \"file_url\" : \"https://cdn.musica.com/tracks/de-acero.mp3\", \"album_id\" : \"alb_101\", \"id\" : \"trk_301\", \"title\" : \"De acero\", \"published_at\" : \"2025-10-16T20:00:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_TRACKS = "/tracks";
    /**
     * GET /tracks : Listado de canciones
     * Devuelve el listado de canciones disponibles en la plataforma.
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "listTracks",
        summary = "Listado de canciones",
        description = "Devuelve el listado de canciones disponibles en la plataforma.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Track.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TracksApi.PATH_LIST_TRACKS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Track>> listTracks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"duration\" : \"03:45\", \"file_url\" : \"https://cdn.musica.com/tracks/de-acero.mp3\", \"album_id\" : \"alb_101\", \"id\" : \"trk_301\", \"title\" : \"De acero\", \"published_at\" : \"2025-10-16T20:00:00Z\" }, { \"duration\" : \"03:45\", \"file_url\" : \"https://cdn.musica.com/tracks/de-acero.mp3\", \"album_id\" : \"alb_101\", \"id\" : \"trk_301\", \"title\" : \"De acero\", \"published_at\" : \"2025-10-16T20:00:00Z\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}